+++ libraries/AP_Declination/examples/AP_Declination_test/AP_Declination_test.cpp
@@ -5,7 +5,6 @@
 #include <AP_HAL/AP_HAL.h>
 #include <AP_Math/AP_Math.h>
 #include <AP_Declination/AP_Declination.h>
-#include <AP_Buffer/AP_Buffer.h>
 #include <Filter/Filter.h>
 
 void setup();
@@ -60,8 +59,8 @@ static float get_declination(float lat, float lon)
     lat = constrain_float(lat, -90, 90);
     lon = constrain_float(lon, -180, 180);
 
-    const int16_t latmin = floor(lat / 5) * 5;
-    const int16_t lonmin = floor(lon / 5) * 5;
+    const int16_t latmin = floorf(lat / 5) * 5;
+    const int16_t lonmin = floorf(lon / 5) * 5;
 
     const uint8_t latmin_index = (90 + latmin) / 5;
     const uint8_t lonmin_index = (180 + lonmin) / 5;
